% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lifecycle_activities.R
\name{check_lifecycle_activities}
\alias{check_lifecycle_activities}
\title{Title}
\usage{
check_lifecycle_activities(construction_object)
}
\arguments{
\item{construction_object}{}
}
\description{
Title
}
