% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timestamps.R
\name{check_timestamps}
\alias{check_timestamps}
\title{Title}
\usage{
check_timestamps(construction_object)
}
\arguments{
\item{construction_object}{[list] A construction object for log building. Access with `.construction_object` from .GlobalEnv.}
}
\description{
Title
}
