% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decide_type.R
\name{decide_type}
\alias{decide_type}
\title{Title}
\usage{
decide_type(construction_object)
}
\arguments{
\item{construction_object}{[list] A construction object for log building. Access with `.construction_object` from .GlobalEnv.}
}
\description{
Title
}
