% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_timestamps.R
\name{prepare_timestamps}
\alias{prepare_timestamps}
\title{Title}
\usage{
prepare_timestamps(construction_object)
}
\arguments{
\item{construction_object}{[list] A construction object for log building. Access with `.construction_object` from .GlobalEnv.}
}
\description{
Title
}
