% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_lifecycle_activities.R
\name{recode_lifecycles_activities}
\alias{recode_lifecycles_activities}
\title{Title}
\usage{
recode_lifecycles_activities(construction_object)
}
\arguments{
\item{construction_object}{}
}
\description{
Title
}
