% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_lifecycle_events.R
\name{recode_lifecycles_events}
\alias{recode_lifecycles_events}
\title{Title}
\usage{
recode_lifecycles_events(construction_object)
}
\arguments{
\item{construction_object}{}
}
\description{
Title
}
