% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_log.R
\name{save_log}
\alias{save_log}
\title{Title}
\usage{
save_log(construction_object)
}
\arguments{
\item{construction_object}{[list] A construction object for log building. Access with `.construction_object` from .GlobalEnv.}
}
\description{
Title
}
