% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_activity_instance_id.R
\name{select_activity_instance}
\alias{select_activity_instance}
\title{Title}
\usage{
select_activity_instance(construction_object)
}
\arguments{
\item{construction_object}{}
}
\description{
Title
}
