% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_lifecycle.R
\name{select_lifecycle}
\alias{select_lifecycle}
\title{Title}
\usage{
select_lifecycle(construction_object)
}
\arguments{
\item{construction_object}{[list] A construction object for log building. Access with `.construction_object` from .GlobalEnv.}
}
\description{
Title
}
