% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_log_identifiers.R
\name{select_log_identifiers}
\alias{select_log_identifiers}
\title{Title}
\usage{
select_log_identifiers(construction_object)
}
\arguments{
\item{construction_object}{[list] A construction object for log building. Access with `.construction_object` from .GlobalEnv.}
}
\description{
Title
}
