% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_timestamps.R
\name{select_timestamps}
\alias{select_timestamps}
\title{Title}
\usage{
select_timestamps(construction_object, single)
}
\arguments{
\item{construction_object}{[list] A construction object for log building. Access with `.construction_object` from .GlobalEnv.}

\item{single}{[lgl]: Single timestamp per row if `TRUE` for event log, multiple timestamps per row if `FALSE` for activity log.}
}
\description{
Title
}
