% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_timestamps.R
\name{select_timestamps}
\alias{select_timestamps}
\title{Title}
\usage{
select_timestamps(construction_object, single)
}
\arguments{
\item{construction_object}{Construction object}

\item{single}{Single timestamp per row}
}
\description{
Title
}
